Refactor: Ajustado para separação de arquivo de teste
